server/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
coqui-tts==0.26.0
deepspeed==0.16.5
Flask==3.1.1
pillow==11.2.1
pytesseract==0.3.13
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
waitress==3.0.2
Werkzeug==3.1.3